Switch to tf.keras API
