+++ requirements.txt
@@ -8,7 +8,9 @@ Keras-Preprocessing==1.0.5
 Markdown==3.0.1
 numpy==1.15.4
 pkg-resources==0.0.0
+Pillow==5.3.0
 protobuf==3.6.1
+pytest==4.0.1
 scikit-learn==0.20.0
 scipy==1.1.0
 six==1.11.0
