+++ scripts/raw_tf_retrain.py
@@ -1,0 +1,511 @@
+    # Copyright 2015 The TensorFlow Authors. All Rights Reserved.
+# Modifications copyright (C) 2018 Irina Vidal Migallon
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Simple transfer learning with Mobilenet models.
+
+With support for TensorBoard.
+
+This example shows how to take a Mobilenet model trained on
+ImageNet images, and train a new top layer that can recognize other classes of
+images.
+
+The top layer receives as input a 2048-dimensional vector (1001-dimensional for
+Mobilenet) for each image. We train a softmax layer on top of this
+representation. Assuming the softmax layer contains N labels, this corresponds
+to learning N + 2048*N (or 1001*N)  model parameters corresponding to the
+learned biases and weights.
+
+Here's an example, which assumes you have a folder containing class-named
+subfolders, each full of images for each label. The example folder flower_photos
+should have a structure like this:
+
+~/flower_photos/daisy/photo1.jpg
+~/flower_photos/daisy/photo2.jpg
+...
+~/flower_photos/rose/anotherphoto77.jpg
+...
+~/flower_photos/sunflower/somepicture.jpg
+
+The subfolder names are important, since they define what label is applied to
+each image, but the filenames themselves don't matter. Once your images are
+prepared, you can run the training with a command like this:
+
+
+```bash
+bazel build tensorflow/examples/image_retraining:retrain && \
+bazel-bin/tensorflow/examples/image_retraining/retrain \
+    --image_dir ~/flower_photos
+```
+
+Or, if you have a pip installation of tensorflow, `retrain.py` can be run
+without bazel:
+
+```bash
+python tensorflow/examples/image_retraining/retrain.py \
+    --image_dir ~/flower_photos
+```
+
+You can replace the image_dir argument with any folder containing subfolders of
+images. The label for each image is taken from the name of the subfolder it's
+in.
+
+This produces a new model file that can be loaded and run by any TensorFlow
+program, for example the label_image sample code.
+
+By default this script will use a Mobilenet model. For example:
+
+Run floating-point version of mobilenet:
+```bash
+python tensorflow/examples/image_retraining/retrain.py \
+    --image_dir ~/flower_photos --architecture mobilenet_1.0_224
+```
+
+There are 32 different Mobilenet models to choose from, with a variety of file
+size and latency options. The first number can be '1.0', '0.75', '0.50', or
+'0.25' to control the size, and the second controls the input image size, either
+'224', '192', '160', or '128', with smaller sizes running faster. See
+https://research.googleblog.com/2017/06/mobilenets-open-source-models-for.html
+for more information on Mobilenet.
+
+To use with TensorBoard:
+
+By default, this script will log summaries to /tmp/retrain_logs directory
+
+Visualize the summaries with this command:
+
+tensorboard --logdir /tmp/retrain_logs
+
+"""
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+import argparse
+from datetime import datetime
+import os.path
+import sys
+import tensorflow as tf
+
+from tensorflow.python.platform import gfile
+from .data_preparation import prepare_file_system, maybe_download_and_extract, \
+    create_image_lists, create_model_info, add_jpeg_decoding, save_graph_to_file, get_random_cached_bottlenecks, \
+    cache_bottlenecks
+
+
+def variable_summaries(var):
+    """Attach a lot of summaries to a Tensor (for TensorBoard visualization)."""
+    with tf.name_scope("summaries"):
+        mean = tf.reduce_mean(var)
+        tf.summary.scalar("mean", mean)
+        with tf.name_scope("stddev"):
+            stddev = tf.sqrt(tf.reduce_mean(tf.square(var - mean)))
+            tf.summary.scalar("stddev", stddev)
+            # tf.summary.scalar("max", tf.reduce_max(var))
+            # tf.summary.scalar("min", tf.reduce_min(var))
+    tf.summary.histogram("histogram", var)
+
+
+def add_final_training_ops(class_count, final_tensor_name, bottleneck_tensor,
+                           bottleneck_tensor_size, optimizer_name="sgd"):
+    """Adds a new softmax and fully-connected layer for training.
+
+    We need to retrain the top layer to identify our new classes, so this function
+    adds the right operations to the graph, along with some variables to hold the
+    weights, and then sets up all the gradients for the backward pass.
+
+    The set up for the softmax and fully-connected layers is based on:
+    https://www.tensorflow.org/versions/master/tutorials/mnist/beginners/index.html
+
+    Args:
+      class_count: Integer of how many categories of things we're trying to recognize.
+      final_tensor_name: Name string for the new final node that produces results.
+      bottleneck_tensor: The output of the main CNN graph.
+      bottleneck_tensor_size: How many entries in the bottleneck vector.
+      optimizer_name: The name of the optimizer to use during training.
+
+    Returns:
+      The tensors for the training and cross entropy results, and tensors for the
+      bottleneck input and ground truth input.
+    """
+
+    valid_optimizers = ["sgd", "adam"]
+    assert optimizer_name in valid_optimizers,\
+        "Unknown optimizer ({}). Expected: {}".format(optimizer_name,
+                                                      valid_optimizers)
+    with tf.name_scope("input"):
+        bottleneck_input = tf.placeholder_with_default(
+            bottleneck_tensor,
+            shape=[None, bottleneck_tensor_size],
+            name="BottleneckInputPlaceholder")
+
+        ground_truth_input = tf.placeholder(
+            tf.int64, [None], name="GroundTruthInput")
+
+    # Organizing the following ops as `final_training_ops` so they"re easier to see in TensorBoard
+    layer_name = "final_training_ops"
+    with tf.name_scope(layer_name):
+        with tf.name_scope("weights"):
+            initial_value = tf.truncated_normal(
+                [bottleneck_tensor_size, class_count], stddev=0.001)
+            layer_weights = tf.Variable(initial_value, name="final_weights")
+            # variable_summaries(layer_weights)
+        with tf.name_scope("biases"):
+            layer_biases = tf.Variable(tf.zeros([class_count]), name="final_biases")
+            variable_summaries(layer_biases)
+
+        with tf.name_scope("Wx_plus_b"):
+            logits = tf.matmul(bottleneck_input, layer_weights) + layer_biases
+            tf.summary.histogram("pre_activations", logits)
+
+    final_tensor = tf.nn.softmax(logits, name=final_tensor_name)
+
+    tf.summary.histogram("activations", final_tensor)
+
+    with tf.name_scope("cross_entropy"):
+        cross_entropy_mean = tf.losses.sparse_softmax_cross_entropy(
+            labels=ground_truth_input, logits=logits)
+
+    tf.summary.scalar("cross_entropy", cross_entropy_mean)
+
+    with tf.name_scope("train"):
+        if optimizer_name == "adam":
+            optimizer = tf.train.AdamOptimizer(learning_rate=FLAGS.learning_rate)
+        else:
+            optimizer = tf.train.GradientDescentOptimizer(FLAGS.learning_rate)
+
+        train_step = optimizer.minimize(cross_entropy_mean)
+
+    return train_step, cross_entropy_mean, bottleneck_input, ground_truth_input, final_tensor
+
+
+def create_model_graph(model_info):
+    """"Creates a graph from saved GraphDef file and returns a Graph object.
+
+    Args:
+      model_info: Dictionary containing information about the model architecture.
+
+    Returns:
+      Graph holding the trained network, and various tensors we'll be manipulating.
+    """
+    with tf.Graph().as_default() as graph:
+        model_path = os.path.join(FLAGS.model_dir, model_info["model_file_name"])
+        tf.logging.info("Model path: {}".format(model_path))
+        with gfile.FastGFile(model_path, "rb") as f:
+            graph_def = tf.GraphDef()
+            graph_def.ParseFromString(f.read())
+            bottleneck_tensor, resized_input_tensor = (tf.import_graph_def(
+                graph_def,
+                name="",
+                return_elements=[
+                    model_info["bottleneck_tensor_name"],
+                    model_info["resized_input_tensor_name"],
+                ]))
+    return graph, bottleneck_tensor, resized_input_tensor
+
+
+def add_evaluation_step(result_tensor, ground_truth_tensor):
+    """Inserts the operations we need to evaluate the accuracy of our results.
+
+    Args:
+      result_tensor: The new final node that produces results.
+      ground_truth_tensor: The node we feed ground truth data into.
+
+    Returns:
+      Tuple of (evaluation step, prediction).
+    """
+    with tf.name_scope("accuracy"):
+        with tf.name_scope("correct_prediction"):
+            prediction = tf.argmax(result_tensor, 1)
+            correct_prediction = tf.equal(prediction, ground_truth_tensor)
+        with tf.name_scope("accuracy"):
+            evaluation_step = tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
+    tf.summary.scalar("accuracy", evaluation_step)
+    return evaluation_step, prediction
+
+
+def main(_):
+    # Needed to make sure the logging output is visible.
+    # See https://github.com/tensorflow/tensorflow/issues/3047
+    tf.logging.set_verbosity(tf.logging.INFO)
+
+    # Prepare necessary directories that can be used during training
+    prepare_file_system(FLAGS.summaries_dir)
+
+
+    # Gather information about the model architecture we'll be using.
+    model_info = create_model_info(FLAGS.architecture)
+    if not model_info:
+        tf.logging.error("Did not recognize architecture flag")
+        return -1
+
+    # Set up the pre-trained graph.
+    maybe_download_and_extract(model_info["data_url"], FLAGS.model_dir)
+
+    graph, bottleneck_tensor, resized_image_tensor = (
+        create_model_graph(model_info))
+
+    # Look at the folder structure, and create lists of all the images.
+    image_lists = create_image_lists(FLAGS.image_dir, FLAGS.testing_percentage,
+                                     FLAGS.validation_percentage, FLAGS.max_num_images_per_class)
+    class_count = len(image_lists.keys())
+    if class_count == 0:
+        tf.logging.error("No valid folders of images found at {}".format(FLAGS.image_dir))
+        return -1
+    if class_count == 1:
+        tf.logging.error("Only one valid folder of images found at {} - multiple classes are "
+                         "needed for classification.".format(FLAGS.image_dir))
+        return -1
+
+    with tf.Session(graph=graph) as sess:
+        # Set up the image decoding sub-graph.
+        jpeg_data_tensor, decoded_image_tensor = add_jpeg_decoding(
+            model_info["input_width"], model_info["input_height"],
+            model_info["input_depth"], model_info["input_mean"],
+            model_info["input_std"])
+
+        # We'll make sure we've calculated the 'bottleneck' image summaries and
+        # cached them on disk.
+        cache_bottlenecks(sess, image_lists, FLAGS.image_dir,
+                          FLAGS.bottleneck_dir, jpeg_data_tensor,
+                          decoded_image_tensor, resized_image_tensor,
+                          bottleneck_tensor, FLAGS.architecture)
+
+        # Add the new layer that we'll be training.
+        (train_step, cross_entropy, bottleneck_input, ground_truth_input,
+         final_tensor) = add_final_training_ops(
+            len(image_lists.keys()), FLAGS.final_tensor_name, bottleneck_tensor,
+            model_info["bottleneck_tensor_size"], FLAGS.optimizer_name)
+
+        # Create the operations we need to evaluate the accuracy of our new layer.
+        evaluation_step, prediction = add_evaluation_step(
+            final_tensor, ground_truth_input)
+
+        # Merge all the summaries and write them out to the summaries_dir
+        merged = tf.summary.merge_all()
+        import time
+        timestamp = time.strftime("%H%M%S")
+        train_writer = tf.summary.FileWriter("{}-{}/train".format(FLAGS.summaries_dir, timestamp),
+                                             sess.graph)
+
+        validation_writer = tf.summary.FileWriter("{}-{}/validation".format(FLAGS.summaries_dir,
+                                                                            timestamp))
+
+        # Set up all our weights to their initial default values.
+        init = tf.global_variables_initializer()
+        sess.run(init)
+
+        # Run the training for as many cycles as requested on the command line.
+        for i in range(FLAGS.how_many_training_steps):
+            # Get a batch of input bottleneck values, from the cache stored on disk.
+            (train_bottlenecks,
+             train_ground_truth, _) = get_random_cached_bottlenecks(
+                sess, image_lists, FLAGS.train_batch_size, "training",
+                FLAGS.bottleneck_dir, FLAGS.image_dir, jpeg_data_tensor,
+                decoded_image_tensor, resized_image_tensor, bottleneck_tensor,
+                FLAGS.architecture, FLAGS.max_num_images_per_class)
+            # Feed the bottlenecks and ground truth into the graph, and run a training
+            # step. Capture training summaries for TensorBoard with the `merged` op.
+            train_summary, _ = sess.run(
+                [merged, train_step],
+                feed_dict={bottleneck_input: train_bottlenecks,
+                           ground_truth_input: train_ground_truth})
+            train_writer.add_summary(train_summary, i)
+
+            # Every so often, print out how well the graph is training.
+            is_last_step = (i + 1 == FLAGS.how_many_training_steps)
+            if (i % FLAGS.eval_step_interval) == 0 or is_last_step:
+                train_accuracy, cross_entropy_value = sess.run(
+                    [evaluation_step, cross_entropy],
+                    feed_dict={bottleneck_input: train_bottlenecks,
+                               ground_truth_input: train_ground_truth})
+                tf.logging.info("%s: Step %d: Train accuracy = %.1f%%" %
+                                (datetime.now(), i, train_accuracy * 100))
+                tf.logging.info("%s: Step %d: Cross entropy = %f" %
+                                (datetime.now(), i, cross_entropy_value))
+                validation_bottlenecks, validation_ground_truth, _ = (
+                    get_random_cached_bottlenecks(
+                        sess, image_lists, FLAGS.validation_batch_size, "validation",
+                        FLAGS.bottleneck_dir, FLAGS.image_dir, jpeg_data_tensor,
+                        decoded_image_tensor, resized_image_tensor, bottleneck_tensor,
+                        FLAGS.architecture, FLAGS.max_num_images_per_class))
+                # Run a validation step and capture training summaries for TensorBoard
+                # with the `merged` op.
+                validation_summary, validation_accuracy = sess.run(
+                    [merged, evaluation_step],
+                    feed_dict={bottleneck_input: validation_bottlenecks,
+                               ground_truth_input: validation_ground_truth})
+                validation_writer.add_summary(validation_summary, i)
+                tf.logging.info("%s: Step %d: Validation accuracy = %.1f%% " %
+                                (datetime.now(), i, validation_accuracy * 100))
+
+        # We've completed all our training, so run a final test evaluation on
+        # some new images we haven't used before.
+        test_bottlenecks, test_ground_truth, test_filenames = (
+            get_random_cached_bottlenecks(
+                sess, image_lists, FLAGS.test_batch_size, "testing",
+                FLAGS.bottleneck_dir, FLAGS.image_dir, jpeg_data_tensor,
+                decoded_image_tensor, resized_image_tensor, bottleneck_tensor,
+                FLAGS.architecture, FLAGS.max_num_images_per_class))
+        test_accuracy, predictions = sess.run(
+            [evaluation_step, prediction],
+            feed_dict={bottleneck_input: test_bottlenecks,
+                       ground_truth_input: test_ground_truth})
+        tf.logging.info("Final test accuracy = %.1f%% (N=%d)" %
+                        (test_accuracy * 100, len(test_bottlenecks)))
+
+        # Write out the trained graph and labels with the weights stored as
+        # constants.
+        save_graph_to_file(sess, graph, FLAGS.output_graph, FLAGS.final_tensor_name)
+        with gfile.FastGFile(FLAGS.output_labels, "w") as f:
+            f.write("\n".join(image_lists.keys()) + "\n")
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--image_dir",
+        type=str,
+        default="",
+        help="Path to folders of labeled images."
+    )
+    parser.add_argument(
+        "--output_graph",
+        type=str,
+        default="/tmp/output_graph.pb",
+        help="Where to save the trained graph."
+    )
+    parser.add_argument(
+        "--output_labels",
+        type=str,
+        default="/tmp/output_labels.txt",
+        help="Where to save the trained graph's labels."
+    )
+    parser.add_argument(
+        "--summaries_dir",
+        type=str,
+        default="/tmp/retrain_logs",
+        help="Where to save summary logs for TensorBoard."
+    )
+    parser.add_argument(
+        "--how_many_training_steps",
+        type=int,
+        default=200,
+        help="How many training steps to run before ending."
+    )
+    parser.add_argument(
+        "--learning_rate",
+        type=float,
+        default=0.01,
+        help="How large a learning rate to use when training."
+    )
+    parser.add_argument(
+        "--testing_percentage",
+        type=int,
+        default=10,
+        help="What percentage of images to use as a test set."
+    )
+    parser.add_argument(
+        "--validation_percentage",
+        type=int,
+        default=10,
+        help="What percentage of images to use as a validation set."
+    )
+    parser.add_argument(
+        "--eval_step_interval",
+        type=int,
+        default=10,
+        help="How often to evaluate the training results."
+    )
+    parser.add_argument(
+        "--train_batch_size",
+        type=int,
+        default=100,
+        help="How many images to train on at a time."
+    )
+    parser.add_argument(
+        "--test_batch_size",
+        type=int,
+        default=-1,
+        help="""\
+      How many images to test on. This test set is only used once, to evaluate
+      the final accuracy of the model after training completes.
+      A value of -1 causes the entire test set to be used, which leads to more
+      stable results across runs.\
+      """
+    )
+    parser.add_argument(
+        "--validation_batch_size",
+        type=int,
+        default=100,
+        help="""\
+      How many images to use in an evaluation batch. This validation set is
+      used much more often than the test set, and is an early indicator of how
+      accurate the model is during training.
+      A value of -1 causes the entire validation set to be used, which leads to
+      more stable results across training iterations, but may be slower on large
+      training sets.\
+      """
+    )
+    parser.add_argument(
+        "--model_dir",
+        type=str,
+        default="/tmp/imagenet",
+        help="""\
+      Path to classify_image_graph_def.pb,
+      imagenet_synset_to_human_label_map.txt, and
+      imagenet_2012_challenge_label_map_proto.pbtxt.\
+      """
+    )
+    parser.add_argument(
+        "--bottleneck_dir",
+        type=str,
+        default="/tmp/bottleneck",
+        help="Path to cache bottleneck layer values as files."
+    )
+    parser.add_argument(
+        "--final_tensor_name",
+        type=str,
+        default="final_result",
+        help="""\
+      The name of the output classification layer in the retrained graph.\
+      """
+    )
+    parser.add_argument(
+        "--architecture",
+        type=str,
+        default="mobilenet_1.0_224",
+        help="""\
+      Which model architecture to use. For faster or smaller models, choose a MobileNet with the 
+      form 'mobilenet_<parameter size>_<input_size>'. For example, 'mobilenet_1.0_224' will pick 
+      a model that is 17 MB in size and takes 224 pixel input images.\
+      """)
+    parser.add_argument(
+        "--max_num_images_per_class",
+        type=int,
+        default=2**27-1,
+        help="""\
+      The maximum number of images to be allowed for a single class. The default number is huge 
+      even for COCO or ImageNet standards (2**27-1).\
+      """)
+    parser.add_argument(
+        "--optimizer_name",
+        type=str,
+        default="sgd",
+        help="""\
+      The optimizer to use. Accepted values are currently 'adam' or 'sgd'\
+      """)
+
+    FLAGS, unparsed = parser.parse_known_args()
+    tf.app.run(main=main, argv=[sys.argv[0]] + unparsed)
